Renamed config parameter and fixed ical plugin
The `url` config parameter was not set. Instead I used a custom
`siteurl` value. Due this the `site.url` call in `atom.xml` was empty.
So the url was relativ to the domain which leads to:

http://feeds.feedburner.com/january-2014-meetup

I renamed the siteurl parameter which hopefully should match the jerkyll
doku: http://jekyllrb.com/docs/configuration/

Oh I miss static typing ...